RDD: app: engine: refactor reconciler
Refactor the engine reconciler to take the kind as well as the name of the
object that triggered the reconcile (excluding when the kind is App, where
the name is empty).

This also moves the bulk of the reconciler (triggered by engine events) to
a new file.

This is in preparation for adding a second reconcile path for objects of
kind ImagePullRequest; that will live in the same controller because this
has access to the container engine.  In that case the Kind will be used to
determine how to reconcile.
